fixed apollo client issue
